clean up and consolidate REST API code
