--- mathesar/api/db/viewsets/__init__.py
@@ -1,1 +1,0 @@
-from mathesar.api.db.viewsets.data_files import DataFileViewSet # noqa
