--- mathesar/api/ui/viewsets/__init__.py
@@ -1,1 +1,0 @@
-from mathesar.api.ui.viewsets.users import * # noqa
